requirements.txt
ultralytics==8.3.51
opencv-python==4.10.0.82
numpy==1.25.2
streamlit
pillow